feat: better auto mod
Auto mod now ensures that swears are only detected if the word starts with the swear.
Made regex swears use non-capturing groups (they are faster).
Allows swears to have spaces between them.
Pre-process the swear regex so that it is not processed wach time.

This has been tested on all of our know false positives and no longer triggers.