feat: role managment
The bot now tries to make sure user roles are up to date.
This check is called when a user gets xp and when they join.

Fixes prime being strike in the config.
Stores if a user is a steamowner, and prevents them from getting it again.
Fixes an error that occurs if a member isn't cached when one of their messages is deleted.